feat(jangar): add control-plane runs route
