+++ services/jangar/src/components/app-shell.tsx
@@ -80,7 +80,7 @@ const ROOT_LABELS = new Map<string, string>([
   ['/agents', 'Agents'],
   ['/control-plane', 'Control plane'],
   ['/control-plane/implementation-specs', 'Specs'],
-  ['/control-plane/agent-runs', 'Runs'],
+  ['/control-plane/runs', 'Runs'],
   ['/codex', 'Codex'],
   ['/codex/search', 'Search'],
   ['/codex/runs', 'All runs'],
